+++ front/hooks/useLastMessageGroupObserver.ts
@@ -1,7 +1,7 @@
 import { useEffect } from "react";
 
 import type { MessageWithContentFragmentsType } from "@app/components/assistant/conversation/ConversationViewer";
-import { LAST_MESSAGE_GROUP_ID } from "@app/components/assistant/conversation/messages/MessageGroup";
+import { LAST_MESSAGE_GROUP_ID } from "@app/components/assistant/conversation/MessageGroup";
 
 /**
  * A custom hook to observe when the last message group element becomes visible or not.
